softWare/requirements.txt
azure-functions
azure-storage
json5
PyPDF2
reportlab